KeepSD_board - 목록 조회용 pageable 객체 생성 및 limit의 datatype 변경 (int -> long)

+++ keepsd/src/main/java/net/gentledot/keepsd/mapper/board/BoardMapper.java
@@ -14,7 +14,7 @@ public interface BoardMapper {
     @Select("SELECT board_no, title, content, reference, created_at, updated_at \n" +
             "FROM ksd_board \n" +
             "ORDER BY created_at DESC LIMIT #{contentNumber} OFFSET #{pageOffset}")
-    List<Board> findAll(Integer pageOffset, Integer contentNumber);
+    List<Board> findAll(Long pageOffset, Integer contentNumber);
 
     @Select("SELECT board_no, title, content, reference, created_at, updated_at \n" +
             "FROM ksd_board \n" +
